Add package and organized classes
