+++ src/main/java/Duke.java
@@ -1,3 +1,11 @@
+import duke.exception.CommandNotFoundException;
+import duke.exception.IllegalFormatException;
+import duke.exception.InvalidInputException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.Todo;
+
 import java.util.ArrayList;
 import java.util.Scanner;
