--- src/main/java/Todo.java
@@ -1,11 +1,0 @@
-public class Todo extends Task {
-
-    public Todo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
