+++ src/main/java/duke/exception/CommandNotFoundException.java
@@ -1,2 +1,4 @@
+package duke.exception;
+
 public class CommandNotFoundException extends DukeException{
 }
