+++ src/main/java/duke/exception/DukeException.java
@@ -1,2 +1,4 @@
+package duke.exception;
+
 public class DukeException extends Exception{
 }
