+++ src/main/java/duke/exception/IllegalFormatException.java
@@ -1,2 +1,4 @@
+package duke.exception;
+
 public class IllegalFormatException extends DukeException{
 }
