+++ src/main/java/duke/exception/InvalidInputException.java
@@ -1,2 +1,4 @@
+package duke.exception;
+
 public class InvalidInputException extends DukeException{
 }
