+++ src/main/java/duke/task/Deadline.java
@@ -1,0 +1,14 @@
+package duke.task;
+
+public class Deadline extends Task {
+    protected String by;
+
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = by;
+    }
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + "(by: " + by + ")";
+    }
+}
